cleanup one more legacy type
