+++ lib/process-services/src/lib/process-list/components/start-process.component.spec.ts
@@ -358,7 +358,7 @@ describe('StartProcessComponent', () => {
             fixture.detectChanges();
             await fixture.whenStable();
 
-            expect(component.selectedProcessDef.name).not.toBeDefined();
+            expect(component.selectedProcessDef).toBeUndefined();
         });
 
         describe('dropdown', () => {
