+++ lib/process-services/src/lib/process-list/public-api.ts
@@ -29,8 +29,6 @@ export * from './services/process.service';
 export * from './services/process-filter.service';
 
 // models
-export * from './models/process-definition.model';
-export * from './models/process-instance-filter.model';
 export * from './models/process-instance-variable.model';
 
 export * from './process-list.module';
